InsertNameCell + non editable name when shared
